+++ ref-farming/src/actions_of_farm.rs
@@ -1,8 +1,8 @@
 
 use near_sdk::{env, near_bindgen, Promise};
-
+use near_sdk::json_types::{U128};
 use simple_farm::{SimpleFarm, HRSimpleFarmTerms};
-use crate::utils::{gen_farm_id};
+use crate::utils::{gen_farm_id, MIN_SEED_DEPOSIT};
 use crate::errors::*;
 use crate::*;
 
@@ -11,11 +11,13 @@ use crate::*;
 impl Contract {
     /// create farm and pay for its storage fee
     #[payable]
-    pub fn create_simple_farm(&mut self, terms: HRSimpleFarmTerms) -> FarmId {
+    pub fn create_simple_farm(&mut self, terms: HRSimpleFarmTerms, min_deposit: Option<U128>) -> FarmId {
 
         let prev_storage = env::storage_usage();
 
-        let farm_id = self.internal_add_farm(&terms);
+        let min_deposit: u128 = min_deposit.unwrap_or(U128(MIN_SEED_DEPOSIT)).0;
+
+        let farm_id = self.internal_add_farm(&terms, min_deposit);
 
         // Check how much storage cost and refund the left over back.
         let storage_needed = env::storage_usage() - prev_storage;
@@ -38,9 +40,29 @@ impl Contract {
     /// Adds given farm to the vec and returns it's id.
     /// If there is not enough attached balance to cover storage, fails.
     /// If too much attached - refunds it back.
-    fn internal_add_farm(&mut self, terms: &HRSimpleFarmTerms) -> FarmId {
+    fn internal_add_farm(&mut self, terms: &HRSimpleFarmTerms, min_deposit: Balance) -> FarmId {
         
-        let mut farm_seed = self.get_seed_default(&terms.seed_id);
+        // let mut farm_seed = self.get_seed_default(&terms.seed_id, min_deposit);
+        let mut farm_seed: VersionedFarmSeed;
+        if let Some(fs) = self.get_seed_wrapped(&terms.seed_id) {
+            farm_seed = fs;
+            env::log(
+                format!(
+                    "New farm created In seed {}, with existed min_deposit {}",
+                    terms.seed_id, farm_seed.get_ref().min_deposit
+                )
+                .as_bytes(),
+            );
+        } else {
+            farm_seed = VersionedFarmSeed::new(&terms.seed_id, min_deposit);
+            env::log(
+                format!(
+                    "The first farm created In seed {}, with min_deposit {}",
+                    terms.seed_id, farm_seed.get_ref().min_deposit
+                )
+                .as_bytes(),
+            );
+        }
 
         let farm_id: FarmId = gen_farm_id(&terms.seed_id, farm_seed.get_ref().next_index as usize);
