+++ ref-farming/src/utils.rs
@@ -5,6 +5,7 @@ use uint::construct_uint;
 use crate::{SeedId, FarmId};
 use crate::errors::*;
 
+pub const MIN_SEED_DEPOSIT: u128 = 1_000_000_000_000_000_000;
 pub const MAX_ACCOUNT_LENGTH: u128 = 64;
 /// Amount of gas for fungible token transfers.
 pub const GAS_FOR_FT_TRANSFER: Gas = 10_000_000_000_000;
